Make it possible to download data without importing it

+++ bin/replicate-mysql.sh
@@ -25,6 +25,11 @@ else
   aws --profile govuk-integration s3 cp "s3://${bucket}/mysql/$(date '+%Y-%m-%d')/${archive_file}" "${archive_path}"
 fi
 
+if [[ -n "$SKIP_IMPORT" ]]; then
+  echo "Skipping import as \$SKIP_IMPORT is set"
+  exit 0
+fi
+
 echo "stopping running govuk-docker containers..."
 govuk-docker down
