+++ bin/replicate-postgresql.sh
@@ -34,6 +34,11 @@ else
   aws --profile govuk-integration s3 cp "s3://${bucket}/postgres/$(date '+%Y-%m-%d')/${archive_file}" "${archive_path}"
 fi
 
+if [[ -n "$SKIP_IMPORT" ]]; then
+  echo "Skipping import as \$SKIP_IMPORT is set"
+  exit 0
+fi
+
 echo "stopping running govuk-docker containers..."
 govuk-docker down
